更新CardNodeRepository中的权重类型为Integer，重构IGraphServiceImpl以支持权重比较并调整边的源和目标

+++ cardflow-graph-provider/src/main/java/com/qiu/cardflow/graph/repository/CardNodeRepository.java
@@ -76,7 +76,7 @@ public interface CardNodeRepository extends ListCrudRepository<CardNode, String>
        class TagWeightResult {
               String tagName;
 
-              Long weight;
+              Integer weight;
 
        }
